Moving conversor to helpers
